chore/8836 make schema inspector component stateless
