+++ src/studio/src/designer/frontend/packages/schema-editor/src/components/SchemaInspector.tsx
@@ -1,43 +1,18 @@
-import {
-  AppBar,
-  Checkbox,
-  FormControlLabel,
-  Grid,
-  IconButton,
-} from '@material-ui/core';
+import { AppBar } from '@material-ui/core';
 import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { createStyles, makeStyles } from '@material-ui/core/styles';
 import { TabContext, TabList, TabPanel } from '@material-ui/lab';
-import type {
-  ILanguage,
-  ISchemaState,
-  Restriction,
-  UiSchemaItem,
-} from '../types';
+import type { ILanguage, ISchemaState, UiSchemaItem } from '../types';
 import { ObjectKind } from '../types/enums';
-import { PropertyItem } from './PropertyItem';
-import {
-  addEnum,
-  addProperty,
-  addRestriction,
-  deleteEnum,
-  deleteField,
-  deleteProperty,
-  setPropertyName,
-  setRequired,
-  setRestriction,
-  setRestrictionKey,
-} from '../features/editor/schemaEditorSlice';
 import { getUiSchemaItem, splitParentPathAndName } from '../utils/schema';
 import { getTranslation } from '../utils/language';
-import { RestrictionField } from './RestrictionField';
-import { EnumField } from './EnumField';
 import { SchemaTab } from './SchemaTab';
 import { isFieldRequired, isNameInUse } from '../utils/checks';
-import { AddPropertyButton } from './AddPropertyButton';
-import { ItemDataComponent } from './ItemDataComponent';
-import { InlineObject } from './InlineObject';
+import { ItemRestrictionsTab } from './SchemaInspector/ItemRestrictionsTab';
+import { ItemPropertiesTab } from './SchemaInspector/ItemPropertiesTab';
+import { getObjectKind } from '../utils/ui-schema-utils';
+import { ItemFieldsTab } from './SchemaInspector/ItemFieldsTab';
 
 const useStyles = makeStyles(
   createStyles({
@@ -95,12 +70,10 @@ export interface ISchemaInspectorProps {
 
 export const SchemaInspector = (props: ISchemaInspectorProps) => {
   const classes = useStyles();
-  const dispatch = useDispatch();
-
   const [objectKind, setObjectKind] = React.useState<ObjectKind>(
     ObjectKind.Field,
   );
-  const [isRequired, setIsRequired] = React.useState<boolean>(false);
+
   const [tabIndex, setTabIndex] = React.useState('0');
 
   const selectedId = useSelector((state: ISchemaState) =>
@@ -132,71 +105,22 @@ export const SchemaInspector = (props: ISchemaInspectorProps) => {
 
   const uiSchema = useSelector((state: ISchemaState) => state.uiSchema);
 
+  const isRequired = selectedItem
+    ? isFieldRequired(parentItem, selectedItem)
+    : false;
+  const readOnly = selectedItem?.$ref !== undefined;
+
   useEffect(() => {
-    if (selectedItem) {
+    if (itemToDisplay) {
       if (tabIndex === '2' && itemToDisplay?.type !== 'object') {
         setTabIndex('0');
       }
-
-      setIsRequired(isFieldRequired(parentItem, selectedItem));
-
-      if (
-        selectedItem.$ref !== undefined ||
-        selectedItem.items?.$ref !== undefined
-      ) {
-        setObjectKind(ObjectKind.Reference);
-      } else if (selectedItem.combination) {
-        setObjectKind(ObjectKind.Combination);
-      } else {
-        setObjectKind(ObjectKind.Field);
-      }
+      setObjectKind(getObjectKind(itemToDisplay));
     } else {
-      setIsRequired(false);
       setObjectKind(ObjectKind.Field);
       setTabIndex('0');
     }
-  }, [selectedItem, parentItem, tabIndex, itemToDisplay]);
-
-  const readOnly = selectedItem?.$ref !== undefined;
-
-  const onChangeValue = (path: string, value: any, key: string) =>
-    dispatch(
-      setRestriction({
-        path,
-        value: isNaN(value) ? value : +value,
-        key,
-      }),
-    );
-
-  const onChangeKey = (path: string, oldKey: string, newKey: string) => {
-    if (oldKey === newKey) {
-      return;
-    }
-    dispatch(
-      setRestrictionKey({
-        path,
-        oldKey,
-        newKey,
-      }),
-    );
-  };
-
-  const onChangePropertyName = (path: string, value: string) =>
-    dispatch(
-      setPropertyName({
-        path,
-        name: value,
-      }),
-    );
-
-  const onDeleteFieldClick = (path: string, key: string) =>
-    dispatch(deleteField({ path, key }));
-
-  const onDeleteObjectClick = (path: string) =>
-    dispatch(deleteProperty({ path }));
-
-  const onDeleteEnumClick = (path: string, value: string) =>
-    dispatch(deleteEnum({ path, value }));
+  }, [tabIndex, itemToDisplay]);
 
   const checkIsNameInUse = (name: string) =>
     isNameInUse({
@@ -206,147 +130,10 @@ export const SchemaInspector = (props: ISchemaInspectorProps) => {
       name,
     });
 
-  const onChangeEnumValue = (value: string, oldValue?: string) => {
-    if (itemToDisplay) {
-      dispatch(
-        addEnum({
-          path: itemToDisplay.path,
-          value,
-          oldValue,
-        }),
-      );
-    }
-  };
-
-  const dispatchAddProperty = () => {
-    const path = itemToDisplay?.path;
-    if (path) {
-      dispatch(
-        addProperty({
-          path,
-          keepSelection: true,
-        }),
-      );
-    }
-  };
-
-  const onAddPropertyClicked = (event: React.BaseSyntheticEvent) => {
-    event.preventDefault();
-    dispatchAddProperty();
-  };
-
-  const onAddRestrictionClick = (event?: React.BaseSyntheticEvent) => {
-    event?.preventDefault();
-    if (itemToDisplay) {
-      dispatch(
-        addRestriction({
-          path: itemToDisplay.path,
-          key: '',
-          value: '',
-        }),
-      );
-    }
-  };
-
-  const dispatchAddEnum = () => {
-    if (itemToDisplay) {
-      dispatch(
-        addEnum({
-          path: itemToDisplay.path,
-          value: 'value',
-        }),
-      );
-    }
-  };
-
-  const onAddEnumButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault();
-    dispatchAddEnum();
-  };
-  const handleRequiredChanged = (e: any, checked: boolean) => {
-    if (selectedItem) {
-      dispatch(
-        setRequired({
-          path: selectedId,
-          key: selectedItem.displayName,
-          required: checked,
-        }),
-      );
-      setIsRequired(checked);
-    }
-  };
-
-  const renderItemProperties = (item: UiSchemaItem) => {
-    if (!item.properties) return null;
-
-    return item.properties.map((p: UiSchemaItem) => {
-      return (
-        <PropertyItem
-          language={props.language}
-          key={p.path}
-          required={item.required?.includes(p.displayName)}
-          readOnly={readOnly}
-          value={p.displayName}
-          fullPath={p.path}
-          onChangeValue={onChangePropertyName}
-          onDeleteField={onDeleteObjectClick}
-          onEnterKeyPress={dispatchAddProperty}
-        />
-      );
-    });
-  };
-
-  const renderItemRestrictions = (item: UiSchemaItem) =>
-    item.restrictions?.map((field: Restriction) => {
-      if (field.key && field.key.startsWith('@')) {
-        return null;
-      }
-      return (
-        <RestrictionField
-          key={field.key}
-          language={props.language}
-          type={itemToDisplay?.type}
-          value={field.value}
-          keyName={field.key}
-          readOnly={readOnly}
-          path={item.path}
-          onChangeValue={onChangeValue}
-          onChangeKey={onChangeKey}
-          onDeleteField={onDeleteFieldClick}
-          onReturn={onAddRestrictionClick}
-        />
-      );
-    });
-
-  const renderEnums = (item: UiSchemaItem) =>
-    item.enum?.map((value: string) => (
-      <EnumField
-        key={value}
-        language={props.language}
-        path={item.path}
-        fullWidth={true}
-        value={value}
-        onChange={onChangeEnumValue}
-        onDelete={onDeleteEnumClick}
-        onEnterKeyPress={dispatchAddEnum}
-      />
-    ));
-
   const handleTabChange = (event: any, newValue: string) =>
     setTabIndex(newValue);
 
-  if (!selectedId) {
-    return (
-      <div>
-        <p className={classes.noItem} id='no-item-paragraph'>
-          {getTranslation('no_item_selected', props.language)}
-        </p>
-        <hr className={classes.divider} />
-      </div>
-    );
-  }
-
-  return (
+  return selectedId ? (
     <div className={classes.root} data-testid='schema-inspector'>
       <TabContext value={tabIndex}>
         <AppBar position='static' color='default' className={classes.appBar}>
@@ -372,86 +159,38 @@ export const SchemaInspector = (props: ISchemaInspectorProps) => {
           </TabList>
         </AppBar>
         <TabPanel value='0'>
-          {selectedItem?.combinationItem && selectedItem.$ref === undefined ? (
-            <InlineObject item={selectedItem} language={props.language} />
-          ) : (
-            <ItemDataComponent
-              selectedId={selectedId}
-              selectedItem={selectedItem}
-              parentItem={parentItem}
-              objectKind={objectKind}
-              language={props.language}
-              checkIsNameInUse={checkIsNameInUse}
-            />
-          )}
+          <ItemPropertiesTab
+            checkIsNameInUse={checkIsNameInUse}
+            itemToDisplay={itemToDisplay ?? undefined}
+            language={props.language}
+            parentItem={parentItem ?? undefined}
+          />
         </TabPanel>
         <TabPanel value='1'>
-          <Grid container spacing={1} className={classes.gridContainer}>
-            <Grid item xs={12}>
-              <FormControlLabel
-                className={classes.header}
-                control={
-                  <Checkbox
-                    checked={isRequired}
-                    onChange={handleRequiredChanged}
-                    name='checkedRequired'
-                  />
-                }
-                label={getTranslation('required', props.language)}
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <hr className={classes.divider} />
-            </Grid>
-            <Grid item xs={4}>
-              <p>{getTranslation('keyword', props.language)}</p>
-            </Grid>
-            <Grid item xs={1} />
-            <Grid item xs={7}>
-              <p>{getTranslation('value', props.language)}</p>
-            </Grid>
-            {itemToDisplay && renderItemRestrictions(itemToDisplay)}
-            <IconButton
-              id='add-restriction-button'
-              aria-label={getTranslation('add_restriction', props.language)}
-              onClick={onAddRestrictionClick}
-            >
-              <i className='fa fa-plus' />
-              {getTranslation('add_restriction', props.language)}
-            </IconButton>
-            {selectedItem && selectedItem.type !== 'object' && (
-              <>
-                <Grid item xs={12}>
-                  <hr className={classes.divider} />
-                  <p className={classes.header}>
-                    {getTranslation('enum', props.language)}
-                  </p>
-                </Grid>
-                {itemToDisplay && renderEnums(itemToDisplay)}
-                <IconButton
-                  id='add-enum-button'
-                  aria-label={getTranslation('add_enum', props.language)}
-                  onClick={onAddEnumButtonClick}
-                >
-                  <i className='fa fa-plus' />
-                  {getTranslation('add_enum', props.language)}
-                </IconButton>
-              </>
-            )}
-          </Grid>
+          <ItemRestrictionsTab
+            classes={classes}
+            isRequired={isRequired}
+            itemToDisplay={itemToDisplay ?? undefined}
+            language={props.language}
+            readonly={readOnly}
+          />
         </TabPanel>
         <TabPanel value='2'>
-          <Grid container spacing={3} className={classes.gridContainer}>
-            {itemToDisplay && renderItemProperties(itemToDisplay)}
-          </Grid>
-          {!readOnly && (
-            <AddPropertyButton
-              onAddPropertyClick={onAddPropertyClicked}
-              language={props.language}
-            />
-          )}
+          <ItemFieldsTab
+            classes={classes}
+            itemToDisplay={itemToDisplay ?? undefined}
+            language={props.language}
+            readonly={readOnly}
+          />
         </TabPanel>
       </TabContext>
+    </div>
+  ) : (
+    <div>
+      <p className={classes.noItem} id='no-item-paragraph'>
+        {getTranslation('no_item_selected', props.language)}
+      </p>
+      <hr className={classes.divider} />
     </div>
   );
 };
