+++ src/studio/src/designer/frontend/packages/schema-editor/src/utils/ui-schema-utils.ts
@@ -1,0 +1,12 @@
+import { UiSchemaItem } from '../types';
+import { ObjectKind } from '../types/enums';
+
+export function getObjectKind(item?: UiSchemaItem) {
+  if (item?.$ref !== undefined || item?.items?.$ref !== undefined) {
+    return ObjectKind.Reference;
+  } else if (item?.combination) {
+    return ObjectKind.Combination;
+  } else {
+    return ObjectKind.Field;
+  }
+}
